+++ modules/core/src/ship/magazine.ts
@@ -1,7 +1,7 @@
 import { Schema, type } from '@colyseus/schema';
 
 import { DesignState } from './system';
-import { defectible } from '../defectible';
+import { defectible } from './system';
 import { number2Digits } from '../number-field';
 import { range } from '../range';
